+++ R/use_travis.R
@@ -1,0 +1,10 @@
+#' Add travis CI configuration file.
+#'
+#' @param force a logical. 
+#'
+#' @export
+
+use_travis <- function(force = FALSE) {
+  out <- add_file("travis.yml", force = force, pre = ".")
+  invisible(out)
+}