+++ tests/testthat/testthat.R
@@ -1,4 +1,2 @@
 library(testthat)
-library(qcbsRworkshops)
-
 test_check("qcbsRworkshops")